+++ ntex-rt/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [0.4.27] - 2025-03-14
 
+* Add srbiters pings ttl
+
+* Retrieves a list of all arbiters in the system
+
 * Add "neon" runtime support
 
 * Drop glommio support